+++ lib/queued-subscription-server/queued_subscription_request/_db/queued-subscription-request-models.d.ts
@@ -1,2 +1,5 @@
-import { DbModelAndOidScope } from '../../../app/server/init-sequelize';
-export declare function getQueuedSubscriptionRequestDbModelAndOidScope(): DbModelAndOidScope[];
+import { QueuedSubscriptionRequestModel } from '../db/queued-subscription-request.model';
+export declare function getQueuedSubscriptionRequestDbModelAndOidScope(): {
+    scope: string;
+    dbModel: typeof QueuedSubscriptionRequestModel;
+}[];