+++ src/gql/relay/relay-edge-connection.builder.ts
@@ -11,6 +11,14 @@ import { GqlBaseAttribs } from './base-attribs.builder';
 import { PageInfo } from './page-info.type';
 
 /**
+ * @see https://facebook.github.io/relay/graphql/connections.htm
+ *
+ * We have to derive a concrete class for Connections, as the typescript introspection
+ * isnt good enougth with generics ( ie the abstract edges cant be decorated successfully as a
+ * graphQL field)...but we can still pull up common beviours to this abstract
+ * class
+ */
+/**
  * @note `SchemaClass` can be passed in addition to `RelayClass` because polymorphic single-table-inheritence
  * typesafety requires it.
  */