+++ src/queued-subscription-server/queued_subscription_request/_db/queued-subscription-request-models.ts
@@ -1,13 +1,11 @@
-import { DbModelAndOidScope } from '../../../app/server/init-sequelize';
 import { QueuedSubscriptionRequestModel } from '../db/queued-subscription-request.model';
 import { getQueuedSubscriptionRequestScopeName } from '../../inititialize-queued-subscription-relay';
-const queuedSubscriptionRequestModels: DbModelAndOidScope[] = [
-  {
-    scope: getQueuedSubscriptionRequestScopeName(),
-    dbModel: QueuedSubscriptionRequestModel
-  }
-];
 
 export function getQueuedSubscriptionRequestDbModelAndOidScope() {
-  return queuedSubscriptionRequestModels;
+  return [
+    {
+      scope: getQueuedSubscriptionRequestScopeName(),
+      dbModel: QueuedSubscriptionRequestModel
+    }
+  ];
 }